Rewrote post course api for new DTO.

+++ CoursePlanner_Backend/Model/Entities/Course.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public int Credit_Hours { get; set; }
+        public double Credit_Hours { get; set; }
         public String Subject { get; set; }
         public int Course_Number { get; set; }
         public List<Feature> Features { get; set; }
